backend/monitoring/requirements.txt
Django==4.2.5
APScheduler==3.10.4
asgiref==3.7.2
django-apscheduler==0.6.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
pip==23.1.2
pytz==2023.3.post1
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
psycopg==3.1.10
psycopg2==2.9.7